#101: Moved the initializers to their own module. This way it will be easier to find and use.
